Update copy during onboarding (#1830)
* Update copy during onboarding

Quite a few users have pointed out that it was not clear that owners can be changed after the Safe is created. E.g. "- Once I’ve created a wallet, it seems it’s possible to revoke and change a signer, but it wasn’t immediately clear if I can change the threshold."

Co-authored-by: Daniel Sanchez <daniel.sanchez@gnosis.pm>