Pin the action to a hash instead of a tag
Using a tag to pin an action instead of the full sha1 hash has a risk,
"because a tag can be moved or deleted if a bad actor gains access to
the repository storing the action", i.e., the contents of the action can
change. By using the full hash this cannot happen.